Changed i3 workspaces to use variables for easier naming.
Hey man, how's it going?  I haven't seen you on steam or discord in a while (152 days).  I just finished up my first year of college and was looking for a way to contact you and decided to do it through this.  I guess you have been busy with University.  Maybe when you are done with this semester we could hang out or something.